+++ Lab6/Task4/homework.sln
@@ -1,9 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-VisualStudioVersion = 15.0.26730.16
+VisualStudioVersion = 15.0.27004.2008
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "homework", "homework\homework.vcxproj", "{3B411ADE-A781-409A-95D3-7A99A6A034AB}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "MergeSort", "homework\MergeSort.vcxproj", "{32E06055-A130-4C28-9A79-44534896D88B}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -13,14 +13,14 @@ Global
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{3B411ADE-A781-409A-95D3-7A99A6A034AB}.Debug|x64.ActiveCfg = Debug|x64
-		{3B411ADE-A781-409A-95D3-7A99A6A034AB}.Debug|x64.Build.0 = Debug|x64
-		{3B411ADE-A781-409A-95D3-7A99A6A034AB}.Debug|x86.ActiveCfg = Debug|Win32
-		{3B411ADE-A781-409A-95D3-7A99A6A034AB}.Debug|x86.Build.0 = Debug|Win32
-		{3B411ADE-A781-409A-95D3-7A99A6A034AB}.Release|x64.ActiveCfg = Release|x64
-		{3B411ADE-A781-409A-95D3-7A99A6A034AB}.Release|x64.Build.0 = Release|x64
-		{3B411ADE-A781-409A-95D3-7A99A6A034AB}.Release|x86.ActiveCfg = Release|Win32
-		{3B411ADE-A781-409A-95D3-7A99A6A034AB}.Release|x86.Build.0 = Release|Win32
+		{32E06055-A130-4C28-9A79-44534896D88B}.Debug|x64.ActiveCfg = Debug|x64
+		{32E06055-A130-4C28-9A79-44534896D88B}.Debug|x64.Build.0 = Debug|x64
+		{32E06055-A130-4C28-9A79-44534896D88B}.Debug|x86.ActiveCfg = Debug|Win32
+		{32E06055-A130-4C28-9A79-44534896D88B}.Debug|x86.Build.0 = Debug|Win32
+		{32E06055-A130-4C28-9A79-44534896D88B}.Release|x64.ActiveCfg = Release|x64
+		{32E06055-A130-4C28-9A79-44534896D88B}.Release|x64.Build.0 = Release|x64
+		{32E06055-A130-4C28-9A79-44534896D88B}.Release|x86.ActiveCfg = Release|Win32
+		{32E06055-A130-4C28-9A79-44534896D88B}.Release|x86.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
