+++ Lab6/Task4/homework/ConsoleWork.h
@@ -1,0 +1,12 @@
+#pragma once
+#define _CRT_SECURE_NO_WARNINGS
+#include <iostream>
+#include <string>
+#include "List.h"
+#include "ListMergeSort.h"
+
+void startWork(List* &list);
+
+void doWork(List* &list);
+
+void endWork(List* &list);
