+++ Lab6/Task4/homework/ListMergeSort.h
@@ -1,0 +1,4 @@
+#pragma once
+#include "List.h"
+
+void listMergeSort(List* &list, int sortBy, bool asc);
